Lint and fix example apps; Move password related stuff to consumer package since it's not shared; add ludsrequirements to strength check responses
