+++ consumerExampleApp/src/main/java/com/stytch/exampleapp/OAuthViewModel.kt
@@ -98,7 +98,7 @@ class OAuthViewModel(application: Application) : AndroidViewModel(application) {
                     val result = StytchClient.handle(it, 60U)
                     _currentResponse.value = when (result) {
                         is DeeplinkHandledStatus.NotHandled -> result.reason
-                        is DeeplinkHandledStatus.Handled -> result.response.toFriendlyDisplay()
+                        is DeeplinkHandledStatus.Handled -> result.response.result.toFriendlyDisplay()
                         // This only happens for password reset deeplinks
                         is DeeplinkHandledStatus.ManualHandlingRequired -> ""
                     }
